cast dtype of predict time array into H dtype

+++ applications/nico2_lib/src/nico2_lib/predictors/_nmf/_nmf_pred.py
@@ -273,7 +273,7 @@ class NmfPredictor:
             for step in self.preprocessing_steps:
                 x = step(x)
         w_query, _, _ = non_negative_factorization(
-            X=x,
+            X=x.astype(self.h_reference.dtype),
             H=self.h_reference[:, indexer],
             init="custom",
             update_H=False,
